set meted dstance of default weapon of humans to Fists.

+++ RandomRpgOnUnity/Assets/Scripts/GenericRPG/Model/Things/Weapons/Weapon.cs
@@ -7,7 +7,7 @@
 
     public class Weapon
     {
-        public double Range { get; set; }
+        public float Range { get; set; }
         public int Damage { get; set; }
         public string Name { get; set; }
         public WeaponType WeaponType { get; set; }
